New Interact System, Puzzle Redesign
I created a new script and added an interactable system with the possibility of changing it distance, and an easy way to make the object interactable depending on what the object is inteded to do.

In "Level 1" Scene I've recreated some desgin like instead of torches being the key to the puzzle solving I've changed them into illuminating crystals to make a difference between the lights of the room and the lights that is going to be the key to solving the puzzle; I've added a pillar alongside 3 different statues and these statues would reperent the order of the lights that needs to be illuminating as the hints are going to be inside the prison cells.