Puzzle Level 1 Complete
Puzzle System, Puzzle Hints, and Environmental Props have Added

Each Room has different props which indicates the sequences of which lights needs to be on or off, and the current sequence is the following O,X,O,X,O,X,O,O (O is Lights On, X is Lights Off).
As for the Order in which these lights is by following the words that is written near the status that is set in the middle of the Room.